requirements.txt
bleach==6.3.0
certifi==2026.4.22
charset-normalizer==3.4.7
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
filelock==3.29.0
fonttools==4.62.1
fsspec==2026.3.0
idna==3.13
Jinja2==3.1.6
kaggle==1.7.4.5
kiwisolver==1.5.0
MarkupSafe==3.0.3
matplotlib==3.10.9
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
opencv-python==4.13.0.92
packaging==26.0
pillow==12.2.0
polars==1.40.1
polars-runtime-32==1.40.1
protobuf==7.34.1
psutil==7.2.2
pydicom==3.0.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.3
requests==2.33.1
scipy==1.15.3
six==1.17.0
sympy==1.14.0
text-unidecode==1.3
torch==2.12.0.dev20260408+cu128
torchaudio==2.11.0.dev20260407+cu128
torchvision==0.27.0.dev20260407+cu128
tqdm==4.67.3
typing_extensions==4.15.0
ultralytics==8.4.41
ultralytics-thop==2.0.19
urllib3==2.6.3
webencodings==0.5.1
